+++ sso_profile/templates/business_profile/admin-collaborator-list.html
@@ -59,6 +59,7 @@
 			<td class="border-white padding-top-15 padding-bottom-10">Request to become {{ collaboration_request.role|lower }}</td>
 			<td class="border-white padding-top-15 padding-bottom-10 numeric">
 			    <form method="post" enctype="multipart/form-data" class="no-default-margin" novalidate="novalidate">
+                    {% csrf_token %}
 		                <input type="hidden" name="request_key" value="{{ collaboration_request.uuid }}" />
 			        <button type="submit" class="link background-white margin-right-30" name="{{ form.action.name }}" value="{{ form.APPROVE }}">Approve</button>
 			        <button type="submit" class="link background-white" name="{{ form.action.name }}" value="{{ form.DELETE }}">Delete request</button>
