+++ sso_profile/templates/business_profile/admin-invite-collaborator.html
@@ -16,6 +16,7 @@
         <h2 class="heading-large margin-top-0 margin-bottom-15">Invite collaborators</h2>
         <p>Invite collaborators and select the role.</p>
         <form method="post" enctype="multipart/form-data" novalidate="novalidate">
+        {% csrf_token %}
         {{ form }}
 	    <details closed>
 	        <summary>Help with roles</summary>
@@ -49,6 +50,7 @@
 			    <td class="border-white padding-top-15 padding-bottom-10">{{ invite.role|title }}</td>
 			    <td class="border-white padding-top-15 padding-bottom-10 numeric">
 			        <form action="{% url 'sso_profile:business-profile-collaboration-invite-delete' %}" method="post" enctype="multipart/form-data" class="no-default-margin">
+                        {% csrf_token %}
 			            <input type="hidden" name="invite_key" value="{{ invite.uuid }}" />
 						<input type="hidden" name="action" value="delete" />
 				    	<input type="submit" class="link background-white numeric" value="Delete">
