+++ tests/unit/sso_profile/business_profile/test_views.py
@@ -7,7 +7,6 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.forms.forms import NON_FIELD_ERRORS
 from django.test import override_settings
 from django.urls import reverse
-from formtools.wizard.views import normalize_name
 from PIL import Image, ImageDraw
 from requests.exceptions import HTTPError
 
@@ -16,7 +15,7 @@ from directory_constants import urls, user_roles
 from sso.helpers import api_client
 from sso_profile.business_profile import constants, forms, helpers, views
 from sso_profile.urls import SIGNUP_URL
-from ..common.helpers import create_response, submit_step_factory
+from ..common.helpers import create_response
 
 pytestmark = pytest.mark.django_db
 
@@ -29,197 +28,6 @@ def create_test_image(extension):
     image.save(byte_io, extension)
     byte_io.seek(0)
     return byte_io
-
-
-@pytest.fixture
-def company_profile_data():
-    return {
-        'name': 'Cool Company',
-        'is_publishable': True,
-        'expertise_products_services': {},
-        'is_identity_check_message_sent': False,
-        'is_published_find_a_supplier': False,
-        'number': '1234567',
-        'slug': 'cool-company',
-        'created': '2012-06-15T13:45:30.00000Z',
-        'modified': '2019-04-05T06:43:23.00000Z',
-    }
-
-
-@pytest.fixture
-def default_private_case_study(case_study_data):
-    return {
-        **case_study_data[views.BASIC],
-        **case_study_data[views.MEDIA],
-        'image_one': 'https://example.com/image-one.png',
-        'image_two': 'https://example.com/image-two.png',
-        'image_three': 'https://example.com/image-three.png',
-    }
-
-
-@pytest.fixture(autouse=True)
-def mock_case_study_retrieve(default_private_case_study):
-    patch = mock.patch.object(
-        api_client.company, 'case_study_retrieve', return_value=create_response(default_private_case_study)
-    )
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture(autouse=True)
-def mock_case_study_update():
-    patch = mock.patch.object(api_client.company, 'case_study_update', return_value=create_response())
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture(autouse=True)
-def mock_case_study_create():
-    patch = mock.patch.object(api_client.company, 'case_study_create', return_value=create_response(201))
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture(autouse=True)
-def mock_retrieve_company(company_profile_data):
-    response = create_response(company_profile_data)
-    patch = mock.patch.object(api_client.company, 'profile_retrieve', return_value=response)
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture(autouse=True)
-def mock_update_company():
-    patch = mock.patch.object(api_client.company, 'profile_update', return_value=create_response())
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture(autouse=True)
-def mock_retrieve_supplier():
-    patch = mock.patch.object(
-        api_client.supplier,
-        'retrieve_profile',
-        return_value=create_response(
-            {
-                'is_company_owner': True,
-                'role': user_roles.ADMIN,
-            }
-        ),
-    )
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture(autouse=True)
-def mock_collaborator_role_update():
-    patch = mock.patch.object(api_client.company, 'collaborator_role_update', return_value=create_response())
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture(autouse=True)
-def mock_collaborator_list(user):
-    response = create_response(
-        [
-            {'sso_id': user.id, 'role': user_roles.ADMIN, 'company_email': user.email, 'name': 'jim example'},
-            {'sso_id': 1234, 'role': user_roles.EDITOR, 'company_email': 'jim@example.com', 'name': 'bob example'},
-        ]
-    )
-    patch = mock.patch.object(api_client.company, 'collaborator_list', return_value=response)
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture(autouse=True)
-def mock_collaboration_request_list(user):
-    response = create_response(
-        [
-            {'requestor_sso_id': user.id, 'uuid': 1234, 'name': 'jim example', 'accepted': False},
-            {'requestor_sso_id': 1234, 'uuid': 1234, 'name': 'bob example', 'accepted': False},
-        ]
-    )
-    patch = mock.patch.object(api_client.company, 'collaboration_request_list', return_value=response)
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture(autouse=True)
-def mock_collaborator_invite_list(user):
-    response = create_response(
-        [
-            {
-                'uuid': '7b4f5c1d-a299-4cb7-aa8e-261101c643de',
-                'collaborator_email': 'jim@example.com',
-                'company': 1,
-                'requestor': 2,
-                'requestor_email': 'requestor@testme.com',
-                'requestor_sso_id': 3,
-                'accepted': False,
-                'role': user_roles.EDITOR,
-            }
-        ]
-    )
-    patch = mock.patch.object(api_client.company, 'collaborator_invite_list', return_value=response)
-    yield patch.start()
-    patch.stop()
-
-
-@pytest.fixture
-def submit_case_study_create_step(client):
-    return submit_step_factory(
-        client=client, url_name='sso_profile:business-profile-case-study', view_class=views.CaseStudyWizardCreateView
-    )
-
-
-@pytest.fixture
-def submit_case_study_edit_step(client):
-    url_name = 'sso_profile:business-profile-case-study-edit'
-    view_class = views.CaseStudyWizardEditView
-    view_name = normalize_name(view_class.__name__)
-    step_names = iter([name for name, form in view_class.form_list])
-
-    def submit_step(data, step_name=None):
-        step_name = step_name or next(step_names)
-        url = reverse(url_name, kwargs={'step': step_name, 'id': 1})
-        response = client.get(url)
-        assert response.status_code == 200
-        return client.post(
-            url,
-            {view_name + '-current_step': step_name, **{step_name + '-' + key: value for key, value in data.items()}},
-        )
-
-    return submit_step
-
-
-@pytest.fixture
-def case_study_data():
-    return {
-        views.BASIC: {
-            'title': 'Example',
-            'description': 'Great',
-            'short_summary': 'Nice',
-            'sector': 'AEROSPACE',
-            'website': 'http://www.example.com',
-            'keywords': 'good, great',
-        },
-        views.MEDIA: {
-            'testimonial': 'Great',
-            'testimonial_name': 'Neville',
-            'testimonial_job_title': 'Abstract hat maker',
-            'testimonial_company': 'Imaginary hats Ltd',
-            'image_one': SimpleUploadedFile(
-                name='image-one.png', content=create_test_image('png').read(), content_type='image/png'
-            ),
-            'image_two': SimpleUploadedFile(
-                name='image-two.png', content=create_test_image('png').read(), content_type='image/png'
-            ),
-            'image_three': '',
-            'image_one_caption': 'nice image',
-            'image_two_caption': 'thing',
-            'image_three_caption': 'thing',
-        },
-    }
 
 
 def test_find_a_buyer_unauthenticated_enrolment(client):
@@ -333,18 +141,13 @@ def test_edit_page_submmit_success(
     assert mock_update_company.call_args == mock.call(sso_session_id=user.session_id, data=data)
 
 
-def test_publish_not_publishable(client, user, mock_retrieve_company, company_profile_data, mock_get_company_profile):
+def test_publish_not_publishable(
+    client, user, mock_retrieve_company, not_publishable_company_profile_data, mock_get_company_profile
+):
     client.force_login(user)
 
     # Give the user a company, so they are not redirected by @sso_profiles.urls.company_required
-    mock_get_company_profile.return_value = company_profile_data
-
-    mock_retrieve_company.return_value = create_response(
-        {
-            **company_profile_data,
-            'is_publishable': False,
-        }
-    )
+    mock_get_company_profile.return_value = not_publishable_company_profile_data
 
     url = reverse('sso_profile:business-profile-publish')
 
@@ -906,7 +709,8 @@ def test_request_identity_verification_already_sent(
     assert mock_verify_identity_request.call_count == 0
 
 
-def test_collaborator_list(mock_collaborator_list, mock_collaboration_request_list, client, user, settings):
+def test_collaborator_list(mock_collaborator_list, mock_collaboration_request_list, user, client, settings):
+
     client.force_login(user)
     mock_collaborator_list.return_value = create_response([])
     mock_collaboration_request_list.return_value = create_response([])
